# todo-service/src/main/resources/openapi/todo-openapi.yaml
openapi: 3.0.0
info:
  title: Todo Service API
  description: API for managing Todos in the Todo platform.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
paths:
  /todos:
    post:
      summary: Create a new Todo
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoRequest'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
    get:
      summary: List Todos
      operationId: listTodos
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: A paginated list of Todos
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/TodoResponse'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
        '401':
          description: Unauthorized

  /todos/{id}:
    put:
      summary: Update an existing Todo
      operationId: updateTodo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoRequest'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Invalid input
        '404':
          description: Todo not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Cannot modify a DONE Todo

    delete:
      summary: Delete a Todo
      operationId: deleteTodo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Todo deleted successfully
        '404':
          description: Todo not found
        '401':
          description: Unauthorized

  /todos/{id}/status:
    patch:
      summary: Change the status of a Todo
      operationId: changeTodoStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          required: true
          schema:
            type: string
            enum: [TODO, IN_PROGRESS, DONE]
      responses:
        '200':
          description: Todo status changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Invalid status
        '404':
          description: Todo not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Cannot change status of a DONE Todo

components:
  schemas:
    TodoRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 3
        userId:
          type: string
      required:
        - title
        - userId

    TodoResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [TODO, IN_PROGRESS, DONE]
        userId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time